move get_orb functions into orbits module
